+++ cognite/neat/graph/stores/__init__.py
@@ -1,5 +1,5 @@
+from ._base import NeatGraphStore, drop_graph_store_storage
 from .configuration import RdfStoreConfig, RdfStoreType
-from .graph_store import NeatGraphStore, drop_graph_store_storage
 from .rdf_to_graph import rdf_file_to_graph
 
 __all__ = ["NeatGraphStore", "drop_graph_store_storage", "RdfStoreType", "RdfStoreConfig", "rdf_file_to_graph"]
