+++ cognite/neat/graph/stores/_base.py
